+++ maps/test.cub
@@ -1,29 +1,31 @@
-NO ./textures/north.xpm
-                        SO ./textures/south.xpm
-           WE ./textures/west.xpm
-EA	./textures/east.xpm
-EA	./textures/east.xpm
-F                  103, 45, 150
-C                90, 63, 171
-11111111
-10000001   11111
-10000001  100001    
-10000001   11001
+NO ./textures/NORTH.xpm
+                        SO ./textures/SOUTH.xpm
+           WE ./textures/WEST.xpm
+EA	./textures/EAST.xpm
+C                  255,0,0
+F                255,255,255
+EA	./textures/EAST.xpm
+C                  255,0,0
+F                255,255,255
+1111111
+10000001  011111
+10000001   100001    
+10000001  011001
 100000001111100 11111
 100000000000000000001
 100000000000000000001         1
 1000000000000000000001       101
  1000000001110000000001     10001
-  1000000000000000000001   1000001
+  1000000000000000000001 111000001
    10000000000100000001 100000001
-    10000000000000000001 000000001
-    100000000001111000001000000000001
-    100 0000001   1000011100000000001
-    1000000001     111111100000000111
+    1000000000000000000110000010011111
+    10000000000111100000H 000W  0000011
+    100 0000001   10000111000000000011
+    1000000001     1111111000000001111
     111111111             100000011
-                           1111111
-                             10S1
-                              101
+                           1111011
+                             101
+                              111
                               11
 1111111111111111111
 1000000000000000001
