+++ srcs/parsing/parsing.h
@@ -27,5 +27,6 @@ bool	check_line(char *line);
 void	copy_map_utils(t_data *data, int fd, char *line);
 void	free_map(char **line, int i, int fd);
 void	fill_blank(t_data *data, int x, int y);
+int		ft_count_comma(char *line);
 
 #endif
