+++ srcs/render/ft_player_mov.c
@@ -100,41 +100,3 @@ void	ft_player_mov_lr(t_data *data)
 		}
 	}
 }
-
-bool	ft_ver_col_return(char **map, double tempx,
-					double tempy, t_data *data)
-{
-	return (map[(int)(tempx - 0.1)][(int)(tempy - 0.1)] != '1'
-		&& map[(int)(tempx + 0.1)][(int)(tempy + 0.1)] != '1'
-		&& map[(int)tempx][(int)tempy] != '1'
-		&& map[(int)(tempx - 0.1)][(int)(tempy - 0.1)] != ' '
-		&& map[(int)(tempx + 0.1)][(int)(tempy + 0.1)] != ' '
-		&& map[(int)tempx][(int)tempy] != ' ' && (int) tempx >= 0
-		&& (int)(tempx + 0.2) <= data->map->max_height
-		&& (int) tempy >= 0
-		&& (int)(tempy + 0.2) < (int) ft_strlen(map[(int) tempx]) - 1);
-}
-
-bool	ft_ver_col(char **map, double tempx, double tempy, t_data *data)
-{
-	t_door	*tmp;
-
-	tmp = data->head_door;
-	if (map[(int)(tempx - 0.2)][(int)(tempy - 0.2)] == 'H'
-		|| map[(int)(tempx + 0.2)][(int)(tempy + 0.2)] == 'H'
-		|| map[(int)tempx][(int)tempy] == 'H')
-	{
-		while (tmp)
-		{
-			if ((tmp->x_pos == (int)(tempx + 0.2)
-				&& tmp->y_pos == (int)(tempy + 0.2))
-				|| (tmp->x_pos == (int)(tempx + 0.2)
-					&& tmp->y_pos == (int)(tempy + 0.2))
-				|| (tmp->x_pos == (int)(tempx)
-					&& tmp->y_pos == (int)(tempy)))
-				return (tmp->pos == 1);
-			tmp = tmp->next;
-		}
-	}
-	return (ft_ver_col_return(map, tempx, tempy, data));
-}
