+++ teste.cub
@@ -1,4 +1,4 @@
-F 76,76,76
+F 0,0,0 1
 C 99,99,99
 NO ./textures/north.xpm
 SO ./textures/south.xpm
